improve: simplify the code to remove "/sdcard/..5.u.S"

+++ inotify.sh
@@ -1,9 +1,7 @@
 TARGET="/sdcard/..5.u.S"
 
 while true; do
-	rm -rf "${TARGET}"
+	[[ -e "${TARGET}" ]] && rm -rf "${TARGET}"
 	sleep 1
-	if [[ ! -e "${TARGET}" ]]; then
-		break
-	fi
+	[[ ! -e "${TARGET}" ]] && break
 done
